Keep image data in uint8 format on CPU
Save CPU, memory, and Ray object store costs by not casting images to bfloat16 before GPU

(cherry picked from commit c9fa182fc14c9cb86132338a80958198f2f83672)

+++ nemo_rl/algorithms/sft.py
@@ -288,7 +288,7 @@ def validate(
             )
 
             # update multimodal data
-            val_data.update(cat_and_padded.get_multimodal_dict(as_tensors=False, pixel_dtype=torch.bfloat16))
+            val_data.update(cat_and_padded.get_multimodal_dict(as_tensors=False))
             # When running validation with drop_last=False, we might end up with a partial batch.
             # Check if we need to pad the final batch to make it divisible by micro_batch_size * dp_size.
             if val_data.size < val_batch_size:
@@ -451,7 +451,7 @@ def sft_train(
                         }
                     )
                     train_data.update(
-                        cat_and_padded.get_multimodal_dict(as_tensors=False, pixel_dtype=torch.bfloat16)
+                        cat_and_padded.get_multimodal_dict(as_tensors=False)
                     )
 
                 print("▶ Taking a training step...")
